grab targets from template repo
