+++ b_RS_data_acquisition.R
@@ -2,7 +2,7 @@ library(targets)
 library(tarchetypes)
 library(reticulate)
 
-yaml_file <- "nw-poudre-historical-config.yml"
+yaml_file <- "example/example_config.yml"
 
 # MUST READ ---------------------------------------------------------------
 
@@ -25,7 +25,6 @@ tar_source("data_acquisition/py/pySetup.R")
 # Source functions --------------------------------------------------------
 
 tar_source("data_acquisition/src/")
-source_python("data_acquisition/py/gee_functions.py")
 
 
 # Define {targets} workflow -----------------------------------------------
@@ -75,7 +74,8 @@ list(
     name = locs,
     command = locs_save,
     read = read_csv(!!.x),
-    packages = "readr"
+    packages = "readr",
+    error = "null"
   ),
   
   # use location shapefile and configurations to get polygons from NHDPlusv2
@@ -137,15 +137,56 @@ list(
       locs
       polygons
       centers
-      ref_pull_457_DSWE1
-      ref_pull_457_DSWE1a
-      ref_pull_89_DSWE1
-      ref_pull_89_DSWE1a
-      ref_pull_457_DSWE3
-      ref_pull_89_DSWE3
       run_GEE_per_tile(WRS_tiles)
     },
     pattern = map(WRS_tiles),
     packages = "reticulate"
+  ),
+  
+  # wait for all earth engine tasks to be completed
+  tar_target(
+    name = ee_tasks_complete,
+    command = {
+      eeRun
+      source_python("data_acquisition/py/poi_wait_for_completion.py")
+    },
+    packages = "reticulate"
+  ),
+  
+  # download all files
+  tar_target(
+    name = download_files,
+    command = {
+      ee_tasks_complete
+      download_csvs_from_drive(drive_folder_name = yml$proj_folder,
+                               google_email = yml$google_email,
+                               version_identifier = yml$run_date)
+    },
+    packages = c("tidyverse", "googledrive")
+  ),
+  
+  # collate all files
+  tar_target(
+    name = make_collated_data_files,
+    command = {
+      download_files
+      collate_csvs_from_drive(file_prefix = yml$proj, 
+                              version_identifier = yml$run_date)
+    },
+    packages = c('tidyverse', 'feather')
+  ),
+  
+  # and collate the data with metadata
+  tar_target(
+    name = make_files_with_metadata,
+    command = {
+      make_collated_data_files
+      add_metadata(yaml = yml,
+                   file_prefix = yml$proj,
+                   version_identifier = yml$run_date)
+    },
+    packages = c("tidyverse", "feather")
   )
+  
+  
 )
