+++ b_RS_data_acquisition/src/calc_center.R
@@ -16,6 +16,9 @@
 #' 
 #' 
 calc_center <- function(poly, yaml) {
+  if (!dir.exists("data_acquisition/out/")) {
+    dir.create("data_acquisition/out/")
+  }
   if (grepl("center", yaml$extent[1])) {
     # create an empty tibble
     poi_df <- tibble(
@@ -24,9 +27,6 @@ calc_center <- function(poly, yaml) {
       poi_latitude = numeric(),
       poi_dist_m = numeric()
     )
-    # create rowids for proper indexing
-    poly <- poly %>% 
-      rowid_to_column("r_id")
     for (i in 1:length(poly[[1]])) {
       poi_df  <- poi_df %>% add_row()
       # grab one polygon
@@ -35,6 +35,10 @@ calc_center <- function(poly, yaml) {
       # Xiao Yang's code in EE - Yang, Xiao. (2020). Deepest point calculation 
       # for any given polygon using Google Earth Engine JavaScript API 
       # (Version v1). Zenodo. https://doi.org/10.5281/zenodo.4136755
+      # this needs to be in WGS for best results
+      if (yaml$poly_crs != "EPSG:4326" & !is.na(yaml$poly_crs)) {
+        one_wbd <- st_transform(one_wbd, "EPSG:4326")
+      }
       coord_for_UTM <- one_wbd %>% st_coordinates()
       mean_x <- mean(coord_for_UTM[ ,1])
       mean_y <- mean(coord_for_UTM[ ,2])
